Review: send.c; fix old (Heirloom) header continuation..
When deciding wether header follow lines should be joined in the
display the current codebase checks wether a header line ends with
a MIME encoded word, and does so if.
However, it didn't, since the checking loop went forward not
backward (cp++ not --cp).  Really ,)
I've added a TODO note in there, too, since that is a decision
that should be left to the backend anyway.