Implement a localopts and etc. bypass..
I once saw a crash after interrupting a macro like grazy that
performs multiple account switches and performs downloads from
within either account.  I've done that several times in a row
because the network was so weak, until finally anything was fine.
The behaviour of S-nail was then odd once we were there, and
i called `set' and it looked all fine except that one value that
should be there was missing.  I'd `var-inspect'ed the value and
it was not there.  Then i called `set' to set it, and that caused
the crash.

I really don't know what happened, but i think it was because
acmava.c:_localopts was set and then we jumped away, later
re-entering due to account changes and macro execution.
Note that the var_clear_allow_undefined setting was always
affected by this problem.

So, until either there are no more jumps at all, or there is
a signal manager that must be peaked from within executing code to
actually invoke signal actions at defined points, implement yet
another temporary restore facility to overcome a general design
flaw of this codebase.