Drop `defines', add `unaccount', rename `undefine'..
The `defines' command has been united with `define' -- just like
`account' will list all defined accounts when used without
arguments `define' will now do so for macros.

`undefine' is the new name of `undef'.

The new command `unaccount' can be used to delete all given
accounts, in equal spirit to `undefine' and macros.
Both will (now) print a message if it is tried to delete something
non-existent, and `unaccount' will not allow to remove the
currently active account.

XXX It is still possible to delete macros while they execute.
XXX We would need to track is-executing recursion, i.e., reference
XXX counting and delete-later bits to make that safe.
XXX But how to combine that with non-local jumps?