Gavin Troy reported SEGV: partially revert [ac694bb]..
[ac694bb] (Two IMAP tries: fix SEGV and useless loads..,
2014-02-18) shouldn't make it into the release as such, because
the single line that i now revert was experimental (the `useless
loads' part).

I kept it in because a lot of testing showed no problems, and
i still need to inspect why boxes with more than a couple of
hundred mails cause the problem (all problem test boxes joined are
507 at the time of this writing, and cause no problems -- doubling
the box results in an immediate crash).

Thanks to Gavin Troy for reporting this including a NYD listing!