YEAH!  Understood Gaetan and Gavin!  Fix cc-test.sh..
Gaetan tried in vain to make me understand that the test is not
working.  Having spend almost two days, including installation of
a Debian VM and other (not such) silly things, i gave up.
Now Gavin reported it too, and send me back the output of its test
run.

And now: the sensation.

I've got it!

Haha!

So let's strip the introductional '[-- Message' line and don't
include it in the checksum generation.

(I am so god damned stupid.)