main.c: set command line options immediately..
For the unlikely case that someone is sending mails from
within startup files, or lists folders or switches accounts
from in there, most of the settings from the command line
would yet have not been honoured because they will all be
set after the initialization sequence is complete.

Instead of writing a TODO entry to check what happens if
those things happen (or even disallow doing so), ensure
that *all* settings also take place immediately.
It blows the program a bit (uselessly!), but due to our
new way of variable handling it is at least quite cheap.