*colour-pagers* -> *colour-pager* ++ (Tarqi Kazan)..
Tarqi reported that he sees escape sequences.
Well, we do not set LESS anymore since v14.7, it was a hacky
approach since it was set in the parent process -- the real
solution would be a struct process or struct child and a command
manager, then adjust what is necessary in the fork(2)ed child,
best with as few page-ins as possible.  We are far from being
there, however...

So instead we've added a hacky single additional environment-add
argument all through our invoke-$PAGER call chain and now set
LESS or LV dependent on what is in $PAGER if the respective
environment variable is not yet set in the child.

And it doesn't make sense to keep a list of pagers in
*colour-pagers* anymore, so i've changed that to the binary
*colour-pager*.  We now have a full set of if/elif/else/endif
conditionals with string matching, shall a user have special
requirements on wether this shall be true or not.

Note that the *colour-pagers* stuff was buggy anyway since it was
checked *before* *colour-terms* ...  Man.