mime.c: local vars were not static; avoid useless file loads..
If we don't find any mime.types(5) files and don't have builtin
commands we yet uselessly try to initialize over and over again.
Instead require the use to `mimetypes clear' to force reload.
Note this can only happen if there are no builtin mimetypes, so
check that (should be optimized away if impossible).