colalign(): fix header display with tabulators..
Finally got around to fix a miss i introduced with [c78df3f]
at the beginning of may: header bodies with tabulators would
cause the replacement character to be printed in multibyte
environments, because iswprint(3) will return false for \t.
Special case that .(, the real solution will be usage of
S-CText once that is usable.