Pass some MIME information to pipe handlers (Bob Tennent)..
Use the environment to pass some informations from MIME headers
down to installed pipe handlers, including a generated filename
that is guaranteed never to be empty.

This enables automatic saving with the correct filename -- that
is, at least a part of the filename that the pipe handler chooses
to use can now be based upon the filename that the sender of the
message had in mind.