urlcrecry.c:_nrc_init(): fix possible FD leak, error message..
If .netrc is not a regular file or accessible by non-user we yet
would have leaked the file descriptor.
Also we printed "Errors occurred while parsing.." when .netrc
was an empty file, fix that while here.

Document clearly that `netrc clear' needs to be called to cause
another attempt to (re) read .netrc.