sendout.c:_sendbundle_setup_creds(): fix false assert() (Jan Chaloupka)..
Again about the still messy state of myorigin()/myaddrs().
With the new URL and credential scheme i've already introduced the
temporary __sendout_ident bypass hack, but that only comes into
play if *smtp-hostname* is set.  If neither that nor *hostname* is
set, *from* is required to exist as a non-NULL entry.  I have
indeed identified this precondition, but used assert(3) to verify
it's true, which is of course completely stupid given that we talk
about something that comes in from the user!

So today Jan Chaloupka (jchaloup AT redhat DOT com) reported on
nail-devel@ that Heirloom mailx can be crashed by setting *smtp*
in combination with with *from* effectively set to a NULL string.
I first was optimistic, but it turns out S-nail can, too.