TUG 1: store history lines as input..
Temporary User Goodie: history entries do not correspond 1:1 to
user input for commands which use laststring() etc., because the
input lines are modified on the fly.  There is a TODO entry to
make input lines constant (at which time we will have the final
solution, with leading and trailing whitespace stripped etc.), and
i once started a topic branch to already implement it, but stopped
because at that time i didn't have had enough overview about the
call chain (far less functions where `const'antified back then)
and stopped because so many functions actually had to create
a line copy.  Anyway, this is shit.

Because of the extremely annoying bug series, finally implement
a temporary user goodie.