Fix `headers' command documentation+ (Jan Chaloupka)..
Jan Chaloupka forwarded to nail-devel@ a Red Hat bugreport [1]
where a user reported that the `headers' command doesn't work as
indicated.  And indeed it is falsely documented, referring to "+"
and "-" as forward and backward scrolling indicators, whereas
`headers' simply takes a message specification.

While here, reorganize the code a little bit so that the thread
and list code doesn't use c_headers() but instead the new
print_header_group(), since in the future all the commands will
take a structure argument in order to keep the command line
constant.

  [1] <https://bugzilla.redhat.com/show_bug.cgi?id=1131533>