Followup for [309dea0] in imap_cache.c..
One more use of string dope storage for .mb_imap_mailbox.
But because of the ridiculous jumping behaviour of the codebase
and usage of global objects we may leave the mailbox hanging
around with a wrong temporary name used when dequeuing the cache.

*However*: i refrain from adding yet another stack of signal
handling because of this shit, all those problems exist in the
codebase since IMAP has been implemented first in nail (11.0+),
and (again) IMAP will (temporarily) vanish in (S-nail) v15.0.