*quote-fold*: fix assignment order reversal..
that causes faulty line length calculation if a line started with
a lot of (leading) whitespace: we need to assign *after* that
whitespace has been taken over, not before (of course).

While here turn a xxx to a TODO: this whitespace turnover will in
fact normalize any whitespace into ASCII SP, instead of taking
care to preserve the buffer range in the original buffer and
"simply" copy that through to our .qf_currq.s "current quote"
storage.