Allow $USER to be an empty string, don't abort(3)..
We yet abort(3) if the user is not known, which is a pretty heavy
action!  Instead exit(67) instead, the BSD /usr/include/sysexits.h
constant for EX_NOUSER.

Also allow an empty user (and an empty -u argument?  ugh!?) to be
idential to an unset $USER, like this invocations like

  $ USER= LC_ALL=C s-nail ...

become possible again.