sendout.c:__prepare_mta_args(): for now, puthead()..
may not always set hp->h_from :(
This is all real, real shit but unfortunately it cannot be helped
for now (pre v15.0) and so be on the safe side.
Better than crashing (which i never have seen however).
Note: this cannot happen, "-r" is _explicitly_ documented to
either take a valid address or require *from* to be set.
I don't care.