collect(): don't copy to stdout unless interactive..
which results in pretty silly things if e.g. *NAIL_HEAD* is set...
But if we do print, ensure stdout is fflushed.

Also, we do not need to print out an initial header if *editalong*
is set and we startup into an editor directly afterwards.
This improves the visual experience somewhat.