Fix *netrc-lookup* -- exclude ambiguous matches, plus..
Herrjemine!

1.) So far, if .netrc included multiple entries for a single host
then if no *user* was set a rather random entry was selected.

Instead of allowing such ambiguities, require an exact match of
its kind, i.e., if there is an exact HOST match then only one
entry of .netrc may match HOST exactly, and ditto for wildcard
matches.  In order for the entry to be selectable in the *nouser*
case, that is.  Exact and wildcard matches may still coexist:

  machine *.a.com login u1 password p1
  machine smtp.a.com login u2 password p2

"u2" will be selected only for a HOST smtp.a.com and "u1" will
only be chosen for any a.com that is not smtp.a.com.

2.) Another bug but in the same domain was yet that a rather
random password would have been chosen if only the password would
have been looked up in .netrc, because in that case the user name
wouldn't have been taken into account at all -- this may be
desired, but ONLY for machine entries without login
specifications:

  machine *.a.com password PASS
  machine *.a.com login diet password curd

So of course diet/curd may ONLY be chosen for user "diet", and
must of course be chosen for it, but all other users may ONLY get
PASS.