SSL: use struct url; use OXM: gain chains; drop *ssl-v2-allow*..
Move away from char buffers, convert anything as necessary to
be able to pass struct url* all through down the call chain.

This step allows simple usage of the xok_*look()up series instead
of unrolling specialized lookup functions, so let's say goodbye to
ssl_method_string(), imap_use_starttls(), _pop3_use_starttls(),
_pop3_no_apop() and whatever else i have forgotten.

At the same time we gain some more user flexibility: say hello to
*ssl-cert-HOST*, *ssl-key-HOST*, *ssl-method-HOST*,
*ssl-verify-HOST* and whatever else i have forgotten.
I think it is just fine not to require *v15-compat* for these.

The initial cause of this commit however was to rework the way we
select the allowed SSL/TLS protocols, so now we do simply and
exclusively use SSLv23_client_method() followed by
SSL_CTX_set_options() with the corresponding bits set.
Not that it really matters, but since we actually *do* call the
latter function anyway...

  Note:
  from what i remember from the first rewrite this change means
  that the semantic of *ssl-method* changes: whereas before
  "tls1.1" meant that _exactly_ TLSv1.1 will be chosen, now
  *ssl-method* rather defines a minimum treshold of acceptable
  protocol versions.

Drop *ssl-v2-allow* and SSLv2 support as such!
The OpenSSL library also dropped support for SSLv2 in 2014-12.
(Ok, on the [master] branch, but this is the way to go.)

And while here, improve verbosity and messages in general.