fio.c:_globname(): use WRDE_NOCMD if available..
This is part of addressing [1] and prevents some command
substitutions in the expanded string -- at least on those systems
which support flags and / or don't implement this by means of the
shell (the simple way).

  [1] http://seclists.org/oss-sec/2014/q4/1066

    0003. Make wordexp support mandatory.  (No functional change.)

No -- wordexp(3) will be chosen on systems which support it.
All do today, but older won't, and i won't give up the possibility
to use S-nail on those.

    0004. Prevent command execution in the expand function, which is IMHO
    unexpected.  (Not really required with patch 1, and there is still
    information disclosure/DoS potential if this expansion occurs.)  This is
    a historic vulnerability already fixed in the Debian package,
    retroactively assigned CVE-2004-2771:

For those systems which have wordexp(3), yes.
Note that we support MIME and thus that retro thing doesn't really
affect us (since we've MIME-decoded the address, did we?).