Change stats handling: fix crash on Bitrig 1.0 (Peter Hofmann)..
Change message lines/bytes statistic handling as detected via
checking the stream we write (as opposed to header display,
which uses the struct message fields) from off_t to ui64_t.

Drop the line counting mechanism: doing so fixes a crash i have
seen on Bitrig 1.0, but actually the bug has been introduced on
2007-06-16 in commit [670ea1c].  We did search the _input_ buffer
for the newline character we base the line count upon, but since
this commit the search in the input buffer extends for as many
bytes as we have generated for output, which is of course terribly
wrong -- that was correct before (as far as searching in the input
buffer can be correct when we are counting output lines).

So for now change the displayed message to only give the actual
output bytes, which affects `copy', `save' etc.  I won't spend any
effort in changing this before v15.0, when our state carrier is
passed all through the call chain, in which case we can hopefully
increment a line counter whenever we place a newline in the actual
output stream.

  IT IS UNBELIEVABLE THAT I HAVEN'T SEEN A CRASH IN ~30 MONTHS
  REGARDING THIS ISSUE!!
  Especially during the QP tests after the bug report from
  Peter Hofmann...  Grazy.

Let's also credit Peter Hofmann, because *if* i would have not
only tried from 1..11111 but really hard, and not only for writing
but also for reading, then i would have found this bug earlier.
I think he really wants to see his name under the fix of this bug.