*Completely* rework MFT: handling..
this not only improves Mail-Followup-To: handling massively by
anticipating some changes of v15 (though i've extended struct
header to transport flags etc. down for now, because that is the
only object that is accessible the way i like things to be in
general), but also fixes other behaviour: e.g., the first ~p in
compose mode will already print a set *replytoo*, for example, and
some other things of this kind.

I also didn't look at mutt(1) again for this round, but simply
stepped on and tried to achieve something that makes sense for me,
like e.g. "overwriting" a List-Post: with a single-address
Reply-To: when about to "auto-`mlist'" a sending list, which e.g.,
fixes IETF and libpng announce message replies, for example.