Penetrate *expandaddr*: introduce enum expand_addr_check_mode..
It must be said: CVE-2014-7844 really was a shock, actually.
I didn't spend a thought on that.  Too.

So, whereas all the terribly redundancy of all that shit won't go
away for quite some time (i'd say:

  [REPLY / -t / ? -> PARSE/CHECK ->]
  COMPOSE -> PARSE/CHECK ->
  SEND

should be the one and only desirable target, right?) we need to
have some straight checking _today_.  Today.