Fix Mail-Followup-To: in certain Reply-To:/List-Post: combinations..
When "automatically learning" mailing-lists as self-announced via
List-Post: headers we until now honoured a set Reply-To: to
forcefully overwrite List-Post:, which e.g. fixes
ietf-announce@ietf.org that sets Reply-To: to ietf@ietf.org
(announce is only used for announcements).

However, if a user sets Reply-To: then this mechanism fails
because the user address is falsely learned as a non-subscribed
mailing-list, causing Mail-Followup-To: to include ourselfs even
if we are subscribed to the actual mailing-list.

To avoid this problem we henceforth only overwrite a Reply-To:
List-Post: if they share the domain name.