Fix autosort crash in non-thread mode..
how good we now have *folder-hook*s with `localopts' so that
i finally come around to use them :), as in (what i and i has)

  set folder-hook-+sent=fh-sent
  define fh-sent {
    localopts yes
    set autosort=date showlast
  }

AND that it is the new year's eve so that i finally did
a "Fi+sent" to an already rotated and thus empty mailbox, because
that resulted.. in a fine crash.
The bug exists since the introduction of threaded display, in
"[11.2] released 8/15/04":

        } else
                threadroot = &message[0];
        finalize(threadroot);

Since threadroot is not NULL it'll be accessed, but since the
mailbox is empty the storage shouldn't.