Fix %$ *headline* output..
YET ANOTHER BUG INTRODUCED IN THE REVIEW FROM FEBRUARY 2014!
This time in [0d4e934] (Review: cmd1.c; *attrlist*: fix, and
ensure correctness, 2014-02-19) there was some resorting going on
and %$ handling was moved from the bottom of the switch() --
apparantly no "break;" was added (it was missing at first glance),
causing the case: to fall through.