Nonsense. Rewrite that again, drop regex..
Ok i am frustrated for more than 15 years due to that restricted
mime.types(5) syntax which was a step backward by not using file
globs already when it was invented.
So now after all this time i had to opportunity to step over that.
And did, and superceeded, by using regular expressions.

So let's drop that again, it is complete overkill for our purpose
and the chance that the increased complexity is doing anything
good is near zero.

Let's however keep the extended f=FILE syntax for
*mimetypes-load-control* and add a different extension for those
files, including the builtin ones: if lines start with "@" then
this defines an automatic plain-text "pipe handler" for this type.
Pfff.

Also add a new "load" subcommand for `mimetype', drop "deleteall"
("delete" does so by default now) and let "add" and "delete" take
multiple arguments instead.

Note that a followup changeset will change that again, deleting
mimetypes has to be implemented by a new `unmimetype', with "*"
meaning "clear the cache".  The interface has to be straight...
(Not yet for `history' though, but has to come.)