nail.1: USAGE INTRODUCTION: on MTA..
two topics on the Arch Linux Forum [1][2] made me realize that we
blindly assume that people know about the Unix mail system, but
today there are distributions which don't ship with the
system-side (sendmail(1) etc.) so that S-nail(1) a.k.a. mailx(1)
won't work out of the box as described!
More on this has to follow, but start with that.

  [1] https://bbs.archlinux.org/viewtopic.php?id=194254
  [2] https://bbs.archlinux.org/viewtopic.php?pid=1507673

While here, unite the "sending mail from scripts" into the usage
introduction, too.