fio.c: add _shvar_exp(), our homebrew FEXP_NSHELL implementation..
On the long run any _globname() has to vanish, at least for file
name expansion purposes.  However, that is *so* hardwired into the
codebase that this is a hopeless task at the very moment.

FEXP_NSHELL had been added for the doable transition.

But we want $VARIABLE and ${VARIABLE} to work, so do implement our
own little, hopefully safe expander for that.