nail.1, mk-conf.sh: on wordexp(3) and new internal..
Talk about the upcoming shell word expansion restrictions, i.e.,
that it will be replaced by a safer but more restricted internal
version (someone should implement a safer in-C wordexp(3), instead
of going through the sh(1) or even, as on Snow Leopard, perl(1)!
Ok i don't care, S-nail will go the a simple internal thing.).

In the latter improve the error message in case wordexp(3) isn't
found, since sofar it sounds as if we now always go through the
shell.  Well yes we do in a lot of places still, but what should
we do until then?  Install perl(1)?