NOPE! -- rework *expandaddr*=fail, `resend' codeflow..
So on top of all that it must be said that our codeflow is
complete horror.  And we need to make an actual difference is
=fail or not =fail.  And, what is yet marked TODO in resend_msg(),
we want to write DEAD if something is wrong with the addressees,
even if that means that we need to do work even if we already know
that we won't send anything out.

So put all that into one changeset on top of the [cve-2014-7844.2]
topic branch.

Let is_addr_invalid() return a "ternary boolean" si8_t meaning "no
error", "error" (TRU1) and "hard failure" (-TRU1).
Change checkaddrs() and namelist_vaporise_head() prototypes and
the sendout.c internal _sendout_error variable to use si8_t.
Etc..