Add NAME_ADDRSPEC_ISMAIL and EACM_NOALIAS..
Until now anything that is not a file or pipe addressee
will be treated as a netword address, but passing "-Atmail"
as a SMTP receiver is possibly not what is desired.

So further classify "names" also as _ISMAIL (XXX very primitively)
and add a new EACM_NOALIAS check mode for is_addr_invalid(), use
it when calling checkaddrs() / namelist_vaporise_head() etc.
accordingly.

XXX If we would expand system-wide aliases then this would require
XXX further diversification