Change `set' output (for non *bsdcompat*, *bsdset*)..
so that the output could (in theory) be fed back into S-nail again
in order to gain the same settings.
To accomplish this a better way, share code with `varshow'.

XXX All of this is very intermediate in sofar as the entire
XXX codebase should be brought in straight shape: sometimes we
XXX want complete backslash escaping (at least \$, \", \\),
XXX sometimes only \" and \\, and sometimes even only \" (e.g.
XXX in here \033 must be possible for at least $prompt).
XXX Today this is patchwork and i won't claim we are correct.
XXX I.e., getrawlist() at minimum must be in line.  blabla
XXX However, this new behaviour of `set' in particular is fine?