A bit more tolerance for faulty Base64 text input..
a message on tz@iana.org contained completely invalid Base64
encoded text, resulting in nothing but "[Invalid Base64 encoding
ignored]".

The message used lines of 990 bytes, splitted up Base64 tuples,
ended (those splitted) lines with invalid characters ("!") and
started "continuation lines" with whitespace.

OpenSSL failed completely, producing no output, but mutt(1) and
Apple Mail are capable to display the message (being silent on
that content).
The absolute horror is that we cannot compare to the latter two,
because of our absolute line-centred way of dealing with about
anything from A to Z.  This is a long, long known issue but the
healing will come with v15.0, no earlier..

However, do no longer abort a faulty B64_T (as opposed to B64)
decoding, but instead continue.  For the message in question this
results in a readable first quarter, completely messed up 60
percent Unicode output (UTF-8 charset), and a readable last ~20
percent, which is better than nothing.

While here, some style tweaks.