Fix temporary variable assignments (Ypnose)..
Ypnose reported [1] that with v14.8.0 his $PAGER is no longer
imported from the environment.  That is because we now explicitly
set some variables which are silently used with default settings
if not set explicitly, so that users will see those default values
in the output of `set': e.g., in an ArchLinux Forum thread [2]
a user simply didn't know where the path "/usr/bin/sendmail" came
from, not knowing it is intransparent fallback for *sendmail*.

Of course (at least) any variable that POSIX defines to be an
environment variable must be inherited from the environment.
Until we have an all-automatic table-driven approach for this, fix
[a61c07e] (main.c: until automatic, set some more variables..,
2015-04-24) and explicitly import SHELL/LISTER/PAGER if possible,
set the fallbacks only otherwise.

  [1] http://article.gmane.org/gmane.mail.s-nail.user/252
  [2] https://bbs.archlinux.org/viewtopic.php?id=194254