sendout.c:__sendout_ident: fix Message-Id: generation..
The __sendout_ident hack that we unfortunately still need for
carrying around information that other parts of the software
easily have available (sick) was yet NULLd whenever _message_id()
was called.  However, since we still do more silly things and try
to generate a complete message in a given character-set (instead
of only the body), starting from scratch shall that fail, the
_message_id() function will be called whenever we try out
a specific character set: obviously our __sendout_ident hack will
be vanished and gone but for the very first round, resulting in
a different Message-Id: to be generated.  So, even though the
latter kind of Message-Id: is nearer to what the others do, it is
not what we generate if we can do better.
Be more personal.  Always.