mk-conf.sh, INSTALL: truly deal with Solaris +..
The real problem on SunOS/Solaris was that we indeed need stuff
from /usr/xpg4 in order to function: the awk(1) from /usr/bin
cannot printf \xHEX sequences (and would require redirecting STDIN
from /dev/null in cc-test.sh), and the termcap(3) from
/usr/include just works fine in the test program, but in all the
header noise that comes in via nail.h i just cannot get around the
first of the following

  In file included from termcap.c:32:
  /usr/include/term.h:1060: error: field `Ottyb' has incomplete type
  /usr/include/term.h:1061: error: field `Nttyb' has incomplete type
  termcap.c: In function `termcap_init':
  termcap.c:87: warning: passing arg 3 of `tputs' from incompatible pointer type
  termcap.c: In function `termcap_destroy':
  termcap.c:101: warning: passing arg 3 of `tputs' from incompatible pointer type
  *** Error code 1

(whereas the latter i already circumvented via [3bfd273], but
which got reverted again).  So use the ncurses(3) from /usr/xpg4
and all is fine.

So generalize this a bit just in case we need more such
OS-specifics in the future.

Update INSTALL accordingly, and add an example on how to adjust
utility programs from the command line.