Recipient address spec. change: "^\./" -> file (Bob Tennent)..
The recipient address type detection is real magic.
Bob correctly pointed out that it is impossible to save a file in
a name that equals the sender:

  resend user@...
  In 14.7.1, this command saves to a file user.

This is even more mystical given that (POSIX) mail(x) supports
some – imho strange nowadays – commands which have exactly this
semantic, e.g., `Copy':

  In the capitalized form, save the specified messages in a file
  whose name is derived from the author of the message

So yes we do

  if (memchr(addr, '@', agp->ag_slen) == NULL) {

and only if that condition is entered "addr" may be a file.
Given the above plus the fact that we already special treat
a leading "|" (vertical bar) to force the remains to be
interpreted as a pipe specification, i think it is well acceptable
to add the special condition of leading "./" (dot slash) to force
the remains to be treated as a file name.