Fix `if' handling invalid compare expressions..
This happens when the test environment cannot be used to test
false syntax but only for testing correctness of valid ones!

So the error with `if' is that invalid operators like @@ or the
like would be swallowed and be rather random (at least in respect
to what the user would expect).