Fix "pipe-TYPE/SUBTYPE=@"..
It has been broken in [b7b673d] (Furtherly extend mime.types
TYPEMARKER extension.., 2015-03-28), which mysteriously removed
the entire condition check.  A real Freudian error.
So oops, that killed a cat: document that the PIPE=@ text-
enforcing case can also be achieved with a special type-marker and
the `mimetype' command, in a (possibly) more useful context.