mk-conf.sh: ensure order of user given flags etc..
ADDCFLAGS/ADDLDFLAGS would have been joined instead of appended
to the other flags, so that a ADDCFLAGS=-std=c99 would not have
worked no more (if we'd detected -std=c89 as a working flag).

While here, ensure that no flag noise sneaks in: reset plain
CFLAGS/LDFLAGS or the ADD* series dependent on wether we do
(not) use WANT_AUTOCC.

And pass the detected flags on to the compile test series, too.