In *debug* mode, "run SMTP" in our own process..
It is completely useless to fork a child and run nothing in it but
printing some debug messages!
And with the new message queues those messages cannot even be
repeated since they happened in their own address space!

So do the same as for normal MTA interaction if *debug* is set,
and that is NOT to start a child; different to MTA interaction TODO
we will simulate the session still, however. TODO