Completely rewrite dot locking..
I've made this partial in (Move file locking to fio.c, drop
dotlock.c..) in the same topic branch, and this changeset rewrites
all the stuff to be hopefully more safe, to allow for
a privilege-separated dotlock program and to actually reflect the
possibility that the (nail)/Heirloom mailx/S-nail codebase has to
offer, and that is "explicit" system mailboxes via the %: prefix.
(So: string comparison against MAILSPOOL never did it, let aside
any symbolic link etc. complication.)