Address struct name memory usage..
I got a weird spam mail message of 16KB that caused 8 string dope
buffers of 31704 bytes each to be allocated, because of
a References: header line of about 12KB (joined)!

We really do all that very bad.  But all of this will have to be
rewritten, and then it is not at all easy anyway, of course.
Especially bad is that we do not even use this during header
overview, only for list/thread and when replying.
XXX We should treat the thread case differently and have a special
XXX parser / storage that saves only those references that are
XXX taken into account, without wasting memory for the others!

For now don't use alloca() for such things anymore, but normal
heap, join the normal (skinned) name (address) of a struct name
into the same storage as the name structure itself and don't
create the n_fullname for GREF instances, which are always
identical to n_name except that they are enclosed in angle
brackets -- instead place those angle brackets when the GREF names
are written.

With these changes applied we "only" need 4 string dope buffers to
parse over that message.  Well, it halved, but still the overhead
is dramatical!